refactor: expose decision engine in public pkg
