+++ pkg/reload/hasher.go
