+++ pkg/reload/hasher_test.go
