+++ pkg/reload/matcher_test.go
@@ -3,7 +3,7 @@ package reload
 import (
 	"testing"
 
-	"github.com/stakater/Reloader/internal/pkg/config"
+	"github.com/stakater/Reloader/pkg/config"
 )
 
 func TestMatcher_ShouldReload(t *testing.T) {
