+++ pkg/reload/resource_type_test.go
